rename to validateTestAssets
