+++ src/wasm-testing-helpers.ts
@@ -11,7 +11,7 @@ import { loadSchema } from './methods/load-schema.js';
 import { loadInputQuery } from './methods/load-input-query.js';
 import { buildFunction } from './methods/build-function.js';
 import { runFunction } from './methods/run-function.js';
-import { validateFixture } from './methods/validate-fixture.js';
+import { validateTestAssets } from './methods/validate-test-assets.js';
 import { validateInputQuery } from './methods/validate-input-query.js';
 import { validateInputQueryFixtureMatch } from './methods/validate-input-query-fixture-match.js';
 import { validateFixtureInput } from './methods/validate-fixture-input.js';
@@ -24,7 +24,7 @@ export {
   loadInputQuery,
   buildFunction,
   runFunction,
-  validateFixture,
+  validateTestAssets,
   validateInputQuery,
   validateInputQueryFixtureMatch,
   validateFixtureInput,
@@ -35,7 +35,7 @@ export {
 export type { FixtureData } from './methods/load-fixture.js';
 export type { BuildFunctionResult } from './methods/build-function.js';
 export type { RunFunctionResult } from './methods/run-function.js';
-export type { ValidateFixtureOptions, CompleteValidationResult } from './methods/validate-fixture.js';
+export type { ValidateTestAssetsOptions, CompleteValidationResult } from './methods/validate-test-assets.js';
 export type { QueryFixtureMatchResult } from './methods/validate-input-query-fixture-match.js';
 export type { ValidationResult } from './methods/validate-fixture-input.js';
 export type { OutputValidationResult } from './methods/validate-fixture-output.js';
